Ian's solution
Solution to the Restaurant Review coding exercise.  I created this WebApi project using Visual Studio 2019.  It uses SQLite as the database, Dapper ORM and repository pattern for data access.  I tested it using Postman.  I exported my Postman tests and included them in this folder.

The database is pre-populated with User and City data, but no Restaurant or Review data.  I ordered the Postman tests so Restaurants and Reviews are posted first, then reads.  The Startup project should be the WebApi project.